Refractored the codebase
