+++ Dockerfile
@@ -1,0 +1,3 @@
+# syntax = docker/dockerfile:1.4
+
+FROM node
