+++ gate-config.yml
@@ -1,5 +1,7 @@
 api_version: v1 #required
 
+# Policies
+
 # Global settings
 settings: #required
   base_path: /api
