+++ helpers/validateGatewayConfig.ts
@@ -1,0 +1,1 @@
+export function validateGateConfig(gateConfigFile: any) {}
