--- nano.14272.save
@@ -1,1 +1,0 @@
-export class constructor() {}
