+++ package.json
@@ -3,17 +3,14 @@
   "version": "1.0.0",
   "description": "An API Gateway implementation",
   "type": "module",
-  "main": "./dists/index.js",
+  "main": "./dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "compile:watch": "tsc -p . -w",
-    "dev": "nodemon ./dists/index.js",
-    "clean": "rimraf ./dists"
+    "dev": "nodemon ./dist/index.js",
+    "clean": "rimraf ./dist"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/biggaji/APIGate.git"
-  },
+  "repository": "git+https://github.com/biggaji/APIGate.git",
   "keywords": [
     "APIManager",
     "APIGateway",
@@ -42,6 +39,7 @@
     "winston": "^3.11.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.2.0",
     "@types/compression": "^1.7.5",
     "@types/cors": "^2.8.16",
     "@types/express": "^4.17.21",
@@ -54,10 +52,17 @@
     "@types/node": "^20.9.0",
     "@types/node-cache": "^4.2.5",
     "@types/winston": "^2.4.4",
+    "@typescript-eslint/eslint-plugin": "^7.8.0",
+    "@typescript-eslint/parser": "^7.8.0",
+    "eslint": "^9.2.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "globals": "^15.2.0",
     "morgan": "^1.10.0",
     "nodemon": "^3.0.1",
     "rimraf": "^5.0.5",
     "ts-node": "^10.9.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.2.2",
+    "typescript-eslint": "^7.8.0"
   }
 }
